requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
peewee==3.13.1
Pillow==7.0.0
pycparser==2.20
python-decouple==3.3
python-telegram-bot==12.4.2
six==1.14.0
tornado==6.0.4
words==0.0.1
